requirements.txt
wheel
Django
python-socketio
eventlet
django-grappelli
django-debug-toolbar
django-debug-panel
Pillow
redis
python-dotenv
GitPython
celery
